+++ src/test/java/net/minestom/server/collision/EntityBlockPhysicsIntegrationTest.java
@@ -272,6 +272,20 @@ public class EntityBlockPhysicsIntegrationTest {
     }
 
     @Test
+    public void entityPhysicsCheckEdgeClip(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 43, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+
+        entity.setInstance(instance, new Pos(0, 42, 0.7)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(10, 0, 0));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(0.7, 42, 0.7), res.newPosition());
+    }
+
+    @Test
     public void entityPhysicsCheckDiagonal(Env env) {
         var instance = env.createFlatInstance();
         instance.setBlock(1, 43, 1, Block.STONE);
@@ -290,6 +304,25 @@ public class EntityBlockPhysicsIntegrationTest {
         // If x collision is checked first then isFirst will be true
         // If z collision is checked first then isSecond will be true
         assertTrue(isFirst || isSecond);
+    }
+
+    @Test
+    public void entityPhysicsCheckCorner(Env env) {
+        var instance = env.createFlatInstance();
+        for (int i = -2; i <= 2; ++i)
+            for (int j = -2; j <= 2; ++j)
+                instance.loadChunk(i, j).join();
+
+        var entity = new Entity(EntityType.ZOMBIE);
+
+        instance.setBlock(5, 43, -5, Block.STONE);
+
+        entity.setInstance(instance, new Pos(-0.3, 42, -0.3)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(10, 0, -10));
+
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(4.7, 42, -10.3), res.newPosition());
     }
 
     @Test
@@ -326,6 +359,192 @@ public class EntityBlockPhysicsIntegrationTest {
         PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0, -20, 0));
 
         assertEqualsPoint(new Pos(5, 43, 5), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsCheckEdgeClipSmall(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 42, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setInstance(instance, new Pos(0.6999, 42, 0.6999)).join();
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.702, 0, 0.702));
+
+        boolean isFirst = checkPoints(new Pos(1.402, 42, 0.7), res.newPosition());
+        boolean isSecond = checkPoints(new Pos(0.7, 42, 1.402), res.newPosition());
+
+        // First and second are both valid, it depends on the implementation
+        // If x collision is checked first then isFirst will be true
+        // If z collision is checked first then isSecond will be true
+        assertTrue(isFirst || isSecond);
+    }
+
+    // Checks C include all checks for crossing one intermediate block (3 block checks)
+    @Test
+    public void entityPhysicsSmallMoveC0(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 42, 0, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.7, 42, 0.5)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.6, 0, 0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1, 42, 1.1), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC1(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(0, 42, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.5, 42, 0.7)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.6, 0, 0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1.1, 42, 1), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC2(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 42, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.8, 42, 1.3)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.6, 0, -0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1, 42, 0.7), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC3(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(0, 42, 0, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.7, 42, 1.1)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.6, 0, -0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1.3, 42, 1), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC4(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(0, 42, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(1.1, 42, 1.3)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(-0.6, 0, -0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1, 42, 0.7), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC5(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 42, 0, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(1.3, 42, 1.1)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(-0.6, 0, -0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(0.7, 42, 1), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC6(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(0, 42, 0, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(1.1, 42, 0.7)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(-0.6, 0, 0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1, 42, 1.3), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC7(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 42, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(1.3, 42, 0.8)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(-0.6, 0, 0.6));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(0.7, 42, 1), res.newPosition());
+    }
+
+    // Checks CE include checks for crossing two intermediate block (4 block checks)
+    @Test
+    public void entityPhysicsSmallMoveC0E(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 43, 0, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.52, 42.51, 0.5)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.57, 0.57, 0.57));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1.09, 43, 1.07), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC1E(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(0, 43, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.50, 42.51, 0.52)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.57, 0.57, 0.57));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1.07, 43, 1.09), res.newPosition());
+    }
+
+    @Test
+    public void entityPhysicsSmallMoveC2E(Env env) {
+        var instance = env.createFlatInstance();
+        instance.setBlock(1, 43, 1, Block.STONE);
+
+        var entity = new Entity(EntityType.ZOMBIE);
+        entity.setBoundingBox(BoundingBox.ZERO);
+
+        entity.setInstance(instance, new Pos(0.51, 42.50, 0.51)).join();
+        assertEquals(instance, entity.getInstance());
+
+        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.57, 0.57, 0.57));
+        EntityBlockPhysicsIntegrationTest.assertEqualsPoint(new Pos(1.08, 43, 1.08), res.newPosition());
     }
 
     @Test
@@ -383,25 +602,6 @@ public class EntityBlockPhysicsIntegrationTest {
 
         assertEquals(new Pos(11, 0, 0), sweepResultFinal.collidedShapePosition);
         assertEquals(sweepResultFinal.collidedShape, bb);
-    }
-
-    @Test
-    public void entityPhysicsCheckEdgeClipSmall(Env env) {
-        var instance = env.createFlatInstance();
-        instance.setBlock(1, 42, 1, Block.STONE);
-
-        var entity = new Entity(EntityType.ZOMBIE);
-        entity.setInstance(instance, new Pos(0.6999, 42, 0.6999)).join();
-
-        PhysicsResult res = CollisionUtils.handlePhysics(entity, new Vec(0.702, 0, 0.702));
-
-        boolean isFirst = checkPoints(new Pos(1.402, 42, 0.7), res.newPosition());
-        boolean isSecond = checkPoints(new Pos(0.7, 42, 1.402), res.newPosition());
-
-        // First and second are both valid, it depends on the implementation
-        // If x collision is checked first then isFirst will be true
-        // If z collision is checked first then isSecond will be true
-        assertTrue(isFirst || isSecond);
     }
 
     @Test
